# bls-server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BLS-KOX API",
    "version": "1.0.0",
    "description": "BLS-KOX 多租户后台管理系统 API 文档\n\n## 认证\n在请求头添加 `Authorization: Bearer <token>`\n登录接口获取 token：`POST /api/auth/login`\n\n## 默认账号\n- 账号: `superadmin`  密码: `123456`  租户: `000000`\n- 账号: `admin`       密码: `123456`  租户: `100000`",
    "contact": {
      "name": "BLS-KOX",
      "url": "https://github.com/npcxl/BLS-KOX"
    }
  },
  "servers": [
    {
      "url": "http://localhost:6001",
      "description": "本地开发"
    },
    {
      "url": "/api",
      "description": "Nginx 反代（生产）"
    }
  ],
  "paths": {
...
    "/api/system/config/current": {
      "get": {
        "tags": [
          "config"
        ],
        "summary": "config - 当前配置",
        "operationId": "get_api_system_config_current",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "操作成功"
                    },
                    "data": {
                      "type": "object",
                      "description": "响应数据"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          }
        }
      }
    },
...
    "/api/system/config/public-system": {
      "get": {
        "tags": [
          "config"
        ],
        "summary": "config - 查询 public-system",
        "operationId": "get_api_system_config_public-system",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "操作成功"
                    },
                    "data": {
                      "type": "object",
                      "description": "响应数据"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          }
        }
      }
    },
    "/api/system/config/public-theme": {
      "get": {
        "tags": [
          "config"
        ],
        "summary": "config - 查询 public-theme",
        "operationId": "get_api_system_config_public-theme",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "操作成功"
                    },
                    "data": {
                      "type": "object",
                      "description": "响应数据"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "登录获取的 accessToken"
      }
    }
  },
...
}